RETHINK actions should notify everyone
This action was only notifying the author of the revision but that was
wrong since who plans changes to a commit (RETHINK) is the revision
author.
This not only notifies the wrong person but also breaks the bot when the
action is performed. That's because we previously removed the actor of
the action from the list of recipients (as usual) thus we couldn't find
the revision author (they are always the same person).

Signed-off-by: Mailson Menezes <mailson.menezes@openbossa.org>